Port animation state functions and talk segue to Rust, fixing null extern crash
Ported 8 static-inline C animation state functions (wantTalkingAnim,
haveTalkingAnim, runningIntroAnim, runningTalkingAnim, setRunIntroAnim,
setRunTalkingAnim, setStopTalkingAnim, has_transition_anim) from null
extern "C" declarations to pure Rust implementations that access
COMM_DATA directly. These were static inline in comm.h and never
compiled as symbols, so the linker resolved them to null via
dynamic_lookup, causing SIGSEGV during the Ur-Quan probe encounter.

Also ported c_DoTalkSegue and c_RunTalkSegue to Rust dinput module
using DoInput-based callbacks. Wired run_talk_segue_dinput() into
ffi.rs to replace c_RunTalkSegue FFI calls.

Fixed pre-existing FFI signature mismatches across modules:
- ReleaseDrawable: battle/c_bridge returned void instead of Drawable
- SetAbsFrameIndex: battle/c_bridge and ships/battle_bridge used c_int
  instead of u16 for frame index
- InitOscilloscope/InitSlider: speech_graphics used u32 instead of
  *mut c_void for frame handles

Verified: 3208 tests pass, 9/9 automation proofs pass, sol-probe-scene-v1
exits cleanly (exit 0), Ur-Quan Kzer-Za alien renders correctly with
proper HUD (CAPTAIN ZELNICK, FUEL:10, CREW:50, FEB 17 2155).

+++ rust/src/ships/battle_bridge.rs
@@ -183,7 +183,7 @@ extern "C" {
     fn TFB_Random() -> u32;
 
     // Frame operations (real C functions)
-    fn SetAbsFrameIndex(frame: Frame, index: c_int) -> Frame;
+    fn SetAbsFrameIndex(frame: Frame, index: u16) -> Frame;
     fn IncFrameIndex(frame: Frame) -> Frame;
     fn GetFrameCount(frame: Frame) -> c_int;
 
@@ -403,7 +403,7 @@ pub mod bridge {
     ///
     /// This is an FFI function called from C. The caller must ensure pointers are valid.
     pub unsafe fn set_abs_frame_index(frame: Frame, index: i32) -> Frame {
-        SetAbsFrameIndex(frame, index as c_int)
+        SetAbsFrameIndex(frame, index as u16)
     }
     /// # Safety
     ///
